simulator/requirements.txt
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
bleach==6.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.1.3
executing==2.0.1
fastjsonschema==2.19.0
fiona==1.9.5
fqdn==1.5.1
geopandas==0.14.1
idna==3.4
influxdb-client==1.37.0
isodate==0.6.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jupyter_core==5.5.0
jupyterlab-pygments==0.2.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.2
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.26.2
osmnx==1.8.0
overrides==7.4.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.0
parso==0.8.3
platformdirs==4.0.0
prometheus-client==0.18.0
prompt-toolkit==3.0.41
psutil==5.9.6
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
pyparsing==3.1.1
pyproj==3.6.1
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
rdflib==7.0.0
reactivex==4.0.4
redis==5.0.3
requests==2.31.0
shapely==2.0.2
SPARQLWrapper==2.0.0
sqlparse==0.4.3
tzdata==2022.7
uri-template==1.3.0
urllib3==2.0.4

